Don't compute least upper bounds for expressions in statement positions inside blocks.
This may save huge amount of time (Fixes SI-5862) for complicated lubs. I had to remove the a check in adapt for
the part that transforms <expr> into { <expr>; () } when the expected type is Unit. The reason is in the
code.

As a side effect, we get more warnings for pure expressions in statement positions (see the change in the test file).