Fix SI-5879.
Fix a bug where a key in an immutable hash map
have the corresponding value different in the iteration
than when doing lookup. This use to happen after calling
`merge`.

Fix the order in which a key-value pair appears in
the collision resolution function - the first argument
always comes from the `this` hash map.

Deprecate `merge` in favour of `merged`, as this is
a pure method.
As an added benefit, the syntax for invoking `merge`
is now nicer.