Avoids cyclic reference error in AddInterfaces.
Compile Global or Typers twice in resident mode yielded a CyclicReference error, in phase erasure. The error was due to a cyclic call of isDerivedValue class. Checking I noted that isDerivedValue class is called very often, and mostly unneccessesarily. I tightened teh condition to not force the type of the class if it is a trait or package. This made the CyclicReference go away.

Widened criterion that enables validation of derived value classes. Previous version made two neg tests slip through.