SI-5162 Exclude super.foo from the erasure cast of SI-4283
If the target method is defined in Java, treat the super reference
as an error, otherwise allow it in the knowledge that Scala loosens
the access restrictions on its generated classes.

Moves the test for that bug out of pending-ville. It's sufficient
to place Test in the empty package to exercise the right code paths.