Fix for broken forwarder.
Wow, I wonder what the upper bound is on how long I might have chased my
tail before finding this. I know what the lower bound is, unfortunately.
It's just that usually the answer to a mystery is not "the empty package
class has magically become the root class, but only sometimes."