fixes a typo in macro classloaders
Primary should be here, because a test a few lines up checks whether
primary != "". Hence usage of fallback here is just a typo.

Primary and fallback classpaths were designed for situations when library
classpath are not enough (e.g. for bootstrapping libraries that don't get
onto tool classpaths). Originally I planned to use them to bootstrap
scala-library, however, there's always a starr/locker on a tool classpath,
so we don't need special treatment of this case. After I realized that, I
decided to keep advanced classpath switches, because others might not be as
lucky as we are.