SI-5999 removes the macro context mirror
It is impossible to enforce macro programmers to use this mirror
instead of c.universe.rootMirror, so it has to be let go.

As to packageless symbol loading, which was the sole purpose of
introducing a special mirror for macro contexts, it will be integrated
into the compiler mirror in subsequent commits.