an improvement based on Adriaan's comment
"""
why use asTypeConstructor for a prefix?
shouldn't that be BaseUniverseClass.thisType?

also, why does the prefix need to be changed at all?
I'm sure there's a valid reason, but it would be good to document it.
if you don't need to change the prefix,
use appliedType(tagType.typeConstructor, List(tpe))
"""

Because we need to look for an implicit of type base.Universe # TypeTag[$tpe].
I figured out that this is exactly the internal representation of that type.