evicts calls to reify from our codebase
Until reflection design is stabilized (i.e. 2.10.0 final is released),
it's a good idea to refrain from using reify in our codebase
(either directly in quasiquotes, or indirectly via materialized type tags).

This increases the percentage of changes to reflection that don't require
rebuilding the starr.

The change to build.xml will expose reifications going on during our build
(by printing out their results to the console, so that they bug everyone),
making it easier to spot and fix them.