Want a 25% partest speedup?*
... well then, don't compile twice!

Explanation:
The compilation process in partest happens in 3 stages:
 - scala + java files, all fed to scalac (so the java signatures are
   loaded and the scala bytecode is generated)
 - java files, fed to javac (so the java bytecode is generated)
 - scala files, fed to scalac (so the scala bytecode correctly links
   to the javac-generated bytecode)

While this mechanism is great to have, for simple 1-file scala tests
it's overkill by compiling scala files twice. So I adjusted the compile
procedure to only run the first step if java files are empty, leading
to a 25% partest speedup.*

Also included Seth Tisue's comment about the tests that require the
three-step compilation.

* as measured on test.scaladoc